feat: add ability to remove product
when deleting a product or a variant, soft deletion is performed,
i.e. only the isActive field is changed to false.
This is due to the fact that it will be problematic to keep track of products
that have been added to the cart or already in the order in the future.

Also now if the product is not found,
the isProductFound field is updated to false
and the message that the product is not found is displayed